Massive overhall, placed everything into packages
