+++ docs/README.md
@@ -1,4 +1,4 @@
-# Clanker User Guide
+# Clanker.command.Clanker User Guide
 
 // Update the title above to match the actual product name
