+++ src/main/java/Clanker/task/Deadline.java
@@ -1,3 +1,5 @@
+package Clanker.task;
+
 public class Deadline extends Task {
     protected String Deadline;
