+++ src/main/java/Clanker/task/Event.java
@@ -1,3 +1,6 @@
+package Clanker.task;
+
+
 public class Event extends Task {
     protected String startTime;
     protected String endTime;
