+++ src/main/java/Clanker/task/Todo.java
@@ -1,4 +1,6 @@
-public class Todo extends Task{
+package Clanker.task;
+
+public class Todo extends Task {
     public Todo(String name) {
         super(name);
     }
