+++ text-ui-test/runtest.bat
@@ -16,7 +16,7 @@ IF ERRORLEVEL 1 (
 REM no error here, errorlevel == 0
 
 REM run the program, feed commands from input.txt file and redirect the output to the ACTUAL.TXT
-java -classpath ..\bin Clanker < input.txt > ACTUAL.TXT
+java -classpath ..\bin Clanker.command.Clanker < input.txt > ACTUAL.TXT
 
 REM compare the output to the expected output
 FC ACTUAL.TXT EXPECTED.TXT
